+++ layout/src/main/java/com/vlkan/log4j2/logstash/layout/resolver/MapResolverFactory.java
@@ -17,7 +17,7 @@ public class MapResolverFactory implements EventResolverFactory<MapResolver> {
 
     @Override
     public MapResolver create(EventResolverContext context, String key) {
-        return new MapResolver(context, key);
+        return new MapResolver(key);
     }
 
 }