+++ layout/src/main/java/com/vlkan/log4j2/logstash/layout/LogstashLayoutSerializationContexts.java
@@ -30,6 +30,8 @@ import com.vlkan.log4j2.logstash.layout.util.ByteBufferOutputStream;
 import java.io.IOException;
 import java.io.OutputStream;
 import java.io.StringReader;
+import java.nio.Buffer;
+import java.nio.ByteBuffer;
 import java.util.function.Supplier;
 
 enum LogstashLayoutSerializationContexts {;
@@ -72,7 +74,9 @@ enum LogstashLayoutSerializationContexts {;
 
                 @Override
                 public void reset() {
-                    outputStream.getByteBuffer().clear();
+                    ByteBuffer byteBuffer = outputStream.getByteBuffer();
+                    // noinspection RedundantCast
+                    ((Buffer) byteBuffer).clear();
                 }
 
             };